refactor(svm): move region-specific parsing out of svm.py into HyundaiBlueLinkApiUSA
svm.py now holds only region-agnostic items: the SVMDetails model,
generic parse helpers, and a recursive redact_svm_metadata that covers
every region's response shape (image always; GPS only for logging —
raw_metadata preserves coordinates, matching prior behavior).

parse_svm_response and the USA-shaped redactors move into
HyundaiBlueLinkApiUSA per review (cdnninja): region-specific parsing
belongs in the region class; svm.py must work for all regions.
